+++ db/PostgreSQL/2.x/2.0/src/02.functions-and-logic/finance.get_account_statement.sql
@@ -3,7 +3,7 @@
     _value_date_from        date,
     _value_date_to          date,
     _user_id                integer,
-    _account_id             bigint,
+    _account_id             integer,
     _office_id              integer
 );
 
@@ -12,7 +12,7 @@ CREATE FUNCTION finance.get_account_statement
     _value_date_from        date,
     _value_date_to          date,
     _user_id                integer,
-    _account_id             bigint,
+    _account_id             integer,
     _office_id              integer
 )
 RETURNS TABLE
